fix(wifi): Fixed crash on radio share
When going from being connected to a network to espnow, the radio would take a long time to deinit then crash since the espnow func would try to access the same resource. This was due to a too long mqtt disconnect. A flag has now been added to speed this up by ensuring mqtt_client_disconnect is only called if it is actually connected instead of waiting to connect then disconnecting 10s later.