fix: Battery will now show charging symbol if charging
This is necessary because the battery voltage will always jump when charging leading to inaccurate readings. The LCD battery icon and text will now turn to a charging symbol when charging to indicate charging without misleading the user. When the charge port is removed, a new updated battery reading will be taken.