fix(lcd_settings, gpio): Changed RGB settings now persist
When RGB LED duration and period settings are changed, the new values now persist to NVS for safekeeping as well as are passed to gpio_funcs to be used live. A mutex has also been added to make sure the two period and duration variables are safe to access across tasks.