feat(lcd_bluetooth, bluetooth): Added media controller
The media controller bluetooth option now displays a media controller icon in which it auto connects to the last known device. From there, the user is able to send vol up, vol down, next, back, and pause/play commands. It will also disconnect automatically when the option is exited from.