fix(lcd_boot): Boot page now only goes away when physically cleared
When the boot page shows up explaining user info, it will now only go away when the user clicks the right arrow. This way, you can now turn off the device or reset it without the boot page clearing so the user will be able to see it upon arrival.